requirements.txt
wheel
git+https://github.com/Pycord-Development/pycord
PyNaCl==1.4.0
requests==2.26.0
dnspython==2.1.0
async-timeout==3.0.1
python-dotenv==0.19.2
pillow==8.3.2
arabic-reshaper==2.1.3
pymongo==4.0.1
pymongo[srv]
humanfriendly==10.0
youtube_dl==2021.12.17